+++ app/Http/Controllers/FastEventController.php
@@ -5,12 +5,17 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Requests\FastEventRequest;
 use App\FastEvent;
+use Gate;
 
 
 class FastEventController extends Controller
 {
     public function store(FastEventRequest $request)
     {
+        if(Gate::denies('edit-posts')){
+            return redirect(route('home'));
+        }
+
         $fastEvent = FastEvent::create($request->all());
 
         return response()->json(['created' => $fastEvent->id]);
@@ -18,6 +23,10 @@ class FastEventController extends Controller
 
     public function update(FastEventRequest $request)
     {
+        if(Gate::denies('edit-posts')){
+            return redirect(route('home'));
+        }
+
         $event = FastEvent::where('id', $request->id)->first();
 
         $event->fill($request->all());
@@ -29,6 +38,10 @@ class FastEventController extends Controller
     
     public function destroy(Request $request)
     {
+        if(Gate::denies('delete-posts')){
+            return redirect(route('home'));
+        }
+
         FastEvent::where('id', $request->id)->delete();
 
         return response()->json(true);
